Feature: support for load functions.

+++ include/garlic/providers/rapidjson.h
@@ -19,7 +19,7 @@ namespace garlic::providers::rapidjson {
       ValueType value;
     };
 
-    using difference_type = int;
+    using difference_type = ptrdiff_t;
     using value_type = MemberWrapper;
     using reference = MemberWrapper&;
     using pointer = MemberWrapper*;
